backend/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
assemblyai==0.31.0
blinker==1.8.2
Brotli==1.1.0
certifi==2024.7.4
charset-normalizer==3.3.2
decorator==4.4.2
distro==1.9.0
Django==5.0.6
django-cors-headers==4.4.0
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
inflection==0.5.1
jiter==0.5.0
moviepy==1.0.3
mutagen==1.47.0
numpy==2.0.1
openai==1.40.3
packaging==24.1
pillow==10.4.0
proglog==0.1.10
psycopg2-binary==2.9.9
pycryptodomex==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
setuptools==70.1.1
sniffio==1.3.1
sqlparse==0.5.0
tqdm==4.66.4
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
websockets==12.0
yt-dlp==2024.8.1
